fix: handle concurrent writes on windows
Windows can return EPERM errors when trying to rename temp files to
files that already exist. In our case a file with a given name will
always have the same content so if it's created while we are trying
to also create it, we can reasonably assume it's ok to use.

If we want to be more thorough we could hash the contents of the new
file.